@@ -2,25 +2,27 @@ name: clima_data
 description: A starting point for Dart libraries or applications.
 
 environment:
-  sdk: '>=2.8.1 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
-  flutter:
-    sdk: flutter
-
-  dartz: 0.9.2
-  http: 0.12.2
-  connectivity: 2.0.1
-  riverpod: 0.12.2
-  shared_preferences: ^0.5.12+4
-
   clima_core:
     path: ../clima_core
   clima_domain:
     path: ../clima_domain
 
+  connectivity: ^3.0.3
+  dartz: ^0.10.0-nullsafety.1
+
+  flutter:
+    sdk: flutter
+
+  http: ^0.13.2
+  riverpod: ^0.14.0+3
+  shared_preferences: ^2.0.5
+
 dev_dependencies:
-  lint: ^1.4.0-dev.d210.1
-  test: ^1.14.4
   flutter_test:
     sdk: flutter
+
+  lint: ^1.5.3
+  test: ^1.16.5
