@@ -3,11 +3,11 @@ import 'dart:async';
 import 'package:clima_ui/main.dart';
 import 'package:clima_ui/state_notifiers/forecasts_state_notifier.dart' as f;
 import 'package:clima_ui/state_notifiers/weather_state_notifier.dart';
+import 'package:clima_ui/utilities/failure_snack_bar.dart';
 import 'package:clima_ui/utilities/hooks.dart';
-import 'package:clima_ui/utilities/reusable_widgets.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
-import 'package:flutter_riverpod/all.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_spinkit/flutter_spinkit.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
@@ -20,9 +20,10 @@ class LoadingScreen extends HookWidget {
   Widget build(BuildContext context) {
     final _themeState = context.read(themeStateNotifier);
     final scaffoldKey = useGlobalKey<ScaffoldState>();
-    final weatherStateNotifier = useProvider(weatherStateNotifierProvider);
+    final weatherStateNotifier =
+        useProvider(weatherStateNotifierProvider.notifier);
     final forecastsStateNotifier =
-        useProvider(f.forecastsStateNotifierProvider);
+        useProvider(f.forecastsStateNotifierProvider.notifier);
 
     useEffect(
       () {
@@ -38,26 +39,20 @@ class LoadingScreen extends HookWidget {
 
           final removeListener = weatherStateNotifier.addListener((state) {
             if (state is Error) {
-              scaffoldKey.currentState.removeCurrentSnackBar();
-              scaffoldKey.currentState.showSnackBar(
-                failureSnackbar(
+              showFailureSnackBar(
+                  scaffoldKey: scaffoldKey,
                   failure: state.failure,
-                  onRetry: load,
-                ),
-              );
+                  onRetry: load);
             }
 
             if (state is Loaded) {
               final removeListener =
                   forecastsStateNotifier.addListener((state) {
                 if (state is f.Error) {
-                  scaffoldKey.currentState.removeCurrentSnackBar();
-                  scaffoldKey.currentState.showSnackBar(
-                    failureSnackbar(
+                  showFailureSnackBar(
+                      scaffoldKey: scaffoldKey,
                       failure: state.failure,
-                      onRetry: load,
-                    ),
-                  );
+                      onRetry: load);
                 }
 
                 if (state is f.Loaded) {
