@@ -1,19 +1,11 @@
 import 'package:clima_core/failure.dart';
 import 'package:flutter/material.dart';
 
-SnackBar snackBar({String text, SnackBarAction action, int duration}) {
-  return SnackBar(
-    behavior: SnackBarBehavior.floating,
-    duration: Duration(seconds: duration ?? 2),
-    content: Text(text),
-    action: action,
-  );
-}
-
-SnackBar failureSnackbar({
+void showFailureSnackBar({
   @required Failure failure,
   VoidCallback onRetry,
   int duration,
+  GlobalKey<ScaffoldState> scaffoldKey,
 }) {
   final text = () {
     if (failure is NoConnection) {
@@ -28,12 +20,15 @@ SnackBar failureSnackbar({
       throw ArgumentError('Did not expect $failure');
     }
   }();
-
-  return snackBar(
-    text: text,
-    duration: duration ?? 86400,
-    action: onRetry != null
-        ? SnackBarAction(label: 'Retry', onPressed: onRetry)
-        : null,
+  scaffoldKey.currentState.removeCurrentSnackBar();
+  scaffoldKey.currentState.showSnackBar(
+    SnackBar(
+      behavior: SnackBarBehavior.floating,
+      content: Text(text),
+      duration: Duration(seconds: duration ?? 86400),
+      action: onRetry != null
+          ? SnackBarAction(label: 'Retry', onPressed: onRetry)
+          : null,
+    ),
   );
 }
